+++ src/lyticsContentProvider.ts
@@ -5,7 +5,7 @@ import { StateManager } from './stateManager';
 import { LyticsUri } from './lyticsUri';
 import { SettingsManager } from './settingsManager';
 import lytics = require("lytics-js/dist/lytics");
-import { LyticsAccount } from 'lytics-js/dist/types';
+import { LyticsAccount, TopicUrl, TopicUrlCollection } from 'lytics-js/dist/types';
 import fs = require('fs');
 import { ContentReader } from './contentReader';
 
@@ -48,6 +48,9 @@ export default class LyticsContentProvider implements vscode.TextDocumentContent
             }
             else if (luri.isTableFieldUri && luri.tableName && luri.tableFieldName) {
                 return await this.provideTextDocumentContentForTableFieldInfo(luri.tableName, luri.tableFieldName, account);
+            }
+            else if (luri.isTopicUri && luri.topicLabel) {
+                return await this.provideTextDocumentContentForTopicInfo(luri.topicLabel, account);
             }
             else if (luri.isEntityUri && luri.tableName && luri.tableFieldName && luri.tableFieldValue) {
                 let entity = await this.provideTextDocumentContentForEntity(luri.tableName, luri.tableFieldName, luri.tableFieldValue, account);
@@ -129,6 +132,18 @@ export default class LyticsContentProvider implements vscode.TextDocumentContent
         const client = lytics.getClient(account.apikey!);
         const info = await client.getTableSchemaFieldInfo(tableName, fieldName);
         return Promise.resolve(JSON.stringify(info ? info : {}, null, 4));
+    }
+    private async provideTextDocumentContentForTopicInfo(topicLabel: string, account: LyticsAccount): Promise<string> {
+        const client = lytics.getClient(account.apikey!);
+        const settings = SettingsManager.getLyticsApiSettings();
+        let collection:(TopicUrlCollection | undefined) = undefined;
+        if (settings.maxTopicUrls > 0) {
+            collection = await client.getTopicUrls(topicLabel, settings.maxTopicUrls);
+        }
+        else {
+            collection = await client.getTopicUrls(topicLabel);
+        }
+        return Promise.resolve(JSON.stringify(collection ? collection : {}, null, 4));
     }
     private async provideTextDocumentContentForEntity(tableName: string, fieldName: string, value: string, account: LyticsAccount): Promise<string> {
         const client = lytics.getClient(account.apikey!);